requirements.txt
### what in requirements.txt file, related to my project. Means all requirment 
### from begning to deployment of website. Remember, I have knowledge of 
### mostily all type of coding language, but i prefer java and free software.
#   pip install -r .\Project_readme\requirements.txt

# === Core Framework ===
Flask==3.0.0                  # Lightweight Python web framework
# OR use Django instead:
# Django==5.0.0               # Full-featured Python web framework

# === Templating & Markdown Support ===
Markdown==3.5.1               # Render .md files into HTML
python-markdown-math==0.8     # Support math formulas in Markdown (LaTeX-style)

# === Form Handling & Validation ===
Flask-WTF==1.2.1              # Form validation and CSRF protection
email-validator==2.1.0        # Validates email fields in forms

# === Database & ORM ===
SQLAlchemy==2.0.23            # ORM for Flask (use with SQLite/PostgreSQL)
Flask-SQLAlchemy==3.1.1       # Flask integration for SQLAlchemy
# OR for Django:
# psycopg2-binary==2.9.9      # PostgreSQL connector for Django

# === Authentication ===
Flask-Login==0.6.3            # User session management
Flask-Bcrypt==1.0.1           # Password hashing

# === File Uploads & PDF Handling ===
Flask-Uploads==0.2.1          # Handle file uploads (PDFs, worksheets)
pdfkit==1.0.0                 # Convert HTML to PDF (requires wkhtmltopdf installed)
PyPDF2==3.0.1                 # Manipulate PDF files (merge, split, etc.)

# === Payment Gateway Integration ===
razorpay==1.3.0               # Razorpay SDK for online payments (India-friendly)
# OR Stripe for global use:
# stripe==8.0.0

# === Deployment & Environment ===
gunicorn==21.2.0              # Production WSGI server for deployment
python-dotenv==1.0.0          # Load environment variables from .env file

# === Security & Utilities ===
itsdangerous==2.1.2           # Secure token generation (used by Flask)
Werkzeug==3.0.1               # Flask's core utility library

# === Optional: Admin Dashboard & API ===
Flask-Admin==1.6.1            # Admin interface for managing users/content
Flask-RESTful==0.3.10         # Build REST APIs (if needed)

# === Optional: Testing & Debugging ===
pytest==8.0.0                 # Unit testing framework
coverage==7.4.0               # Code coverage reports

